Refactor dune files
The release profile currently active in the makefile for the
build-static target appears to override the static build thus breaking
the pre-built linux binary. This PR reverts that change.

It also removes the inclusion of the stdc++ library with appears to be
unused by MLang.

Finally this PR adds O3 level optimizations when the static profile is
active, increasing performance.